UtilityMethodsTraitを追加し、cfFormMailerクラスにユーティリティメソッドを統合
